+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
+### Changed
+- added some proper logging
+
 ### Fixed
 - AI responded to all messages starting with help
 - order_reporting does not start multiple progress reporters anymore